+++ engine-macros/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "engine-macros"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["Let"]
 edition = "2021"
 description = "Macros to the Let engine"
