+++ engine/src/game/objects/mod.rs
@@ -1,5 +1,7 @@
 pub mod data;
 pub mod physics;
+pub mod labels;
+use super::camera::{Camera, CameraScaling};
 use super::{materials, AObject, NObject};
 use crate::error::objects::*;
 use crate::error::textures::*;
@@ -80,34 +82,6 @@ pub trait GameObject: Send + Any {
     fn as_any(&self) -> &dyn Any;
     fn collider_handle(&self) -> Option<rapier2d::geometry::ColliderHandle>;
     fn rigidbody_handle(&self) -> Option<rapier2d::dynamics::RigidBodyHandle>;
-}
-
-pub trait Camera: GameObject {
-    fn settings(&self) -> CameraSettings;
-}
-
-#[derive(Clone, Copy)]
-pub struct CameraSettings {
-    pub zoom: f32,
-    pub mode: CameraScaling,
-}
-impl Default for CameraSettings {
-    fn default() -> Self {
-        Self {
-            zoom: 1.0,
-            mode: CameraScaling::Stretch,
-        }
-    }
-}
-impl CameraSettings {
-    pub fn zoom(mut self, zoom: f32) -> Self {
-        self.zoom = zoom;
-        self
-    }
-    pub fn mode(mut self, mode: CameraScaling) -> Self {
-        self.mode = mode;
-        self
-    }
 }
 
 #[derive(Clone)]
