+++ engine/src/game/objects/physics/colliders.rs
@@ -1,5 +1,5 @@
 use super::VHACDParameters;
-use crate::{Data, Transform};
+use crate::{objects::Data, Transform};
 use glam::Vec2;
 use rapier2d::prelude::*;
